# be/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Private Clinic API
  version: 0.17.0
  description: Public contract for the private clinic management system.
servers:
  - url: http://localhost:5000
paths:
...
  /api/v1/clinical/branches:
    get:
      operationId: listClinicalBranches
      summary: List active branches assigned to the signed-in doctor
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Authorized clinical branches.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalBranchListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/encounters:
    get:
      operationId: listClinicalEncounters
      summary: List the doctor's encounters in one branch
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BranchPublicIdQuery'
        - name: status
          in: query
          required: false
          description: Comma-separated encounter statuses.
          schema: { type: string }
      responses:
        '200':
          description: Up to 100 encounters, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalEncounterListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/encounters/{encounterId}:
    get:
      operationId: getClinicalEncounter
      summary: Read a clinical encounter assigned to the doctor
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      responses:
        '200':
          description: Encounter, measurements, diagnoses, orders, results and amendments.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalEncounterResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/encounters/{encounterId}/start:
    post:
      operationId: startClinicalEncounter
      summary: Start a called encounter, or bypass calling for the next eligible ticket with permission and reason
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClinicalStartRequest' }
      responses:
        '200':
          description: Started encounter.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalEncounterResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/encounters/{encounterId}/clinical-notes:
    patch:
      operationId: updateClinicalNotes
      summary: Replace clinical notes on an in-progress encounter
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/EncounterId'
        - $ref: '#/components/parameters/ClinicalIfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClinicalNotesRequest' }
      responses:
        '200':
          description: Updated encounter.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalEncounterResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /api/v1/encounters/{encounterId}/vital-signs:
    post:
      operationId: addClinicalVitalSigns
      summary: Append a vital-sign measurement while the encounter is open
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClinicalVitalSignsRequest' }
      responses:
        '201':
          description: Measurement created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalCommandResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/encounters/{encounterId}/diagnoses:
    post:
      operationId: addClinicalDiagnosis
      summary: Append a diagnosis; at most one can be primary
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClinicalDiagnosisRequest' }
      responses:
        '201':
          description: Diagnosis created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalCommandResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/encounters/{encounterId}/services:
    post:
      operationId: orderClinicalService
      summary: Order an available branch service at its current branch price
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClinicalOrderServiceRequest' }
      responses:
        '201':
          description: Service order created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalCommandResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/clinical/services/{serviceId}/results/finalize:
    post:
      operationId: finalizeClinicalServiceResult
      summary: Create immutable FINAL result version 1 and complete the service
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterServiceId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClinicalFinalizeResultRequest' }
      responses:
        '201':
          description: Result finalized.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalCommandResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/encounters/{encounterId}/complete:
    post:
      operationId: completeClinicalEncounter
      summary: Complete after a primary diagnosis and all required services are resolved
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      responses:
        '200':
          description: Completed encounter.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalEncounterResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/encounters/{encounterId}/sign:
    post:
      operationId: signClinicalEncounter
      summary: Hash and lock the completed canonical clinical record
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      responses:
        '200':
          description: SHA-256 hash attestation stored on the encounter.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalCommandResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/encounters/{encounterId}/release-to-patient:
    post:
      operationId: releaseClinicalEncounterToPatient
      summary: Release a signed clinical record and its final results to linked patient accounts
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      responses:
        '200':
          description: Released encounter; retries return the original release metadata.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalEncounterResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/encounters/{encounterId}/amendments:
    post:
      operationId: amendClinicalEncounter
      summary: Append a hash-chained amendment to a signed record
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/EncounterId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClinicalAmendmentRequest' }
      responses:
        '201':
          description: Amendment appended.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClinicalCommandResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/patient/clinical-records:
    get:
      operationId: listPatientClinicalRecords
      summary: List signed records released to an actively linked patient profile
      security: [{ bearerAuth: [] }]
      parameters:
        - name: patientPublicId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Released clinical history, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientClinicalRecordListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/patient/clinical-records/{encounterId}:
    get:
      operationId: getPatientClinicalRecord
      summary: Read one released record through an active patient or guardian link
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/EncounterId'
        - name: patientPublicId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Released notes, diagnoses, results, amendments and integrity hash.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PatientClinicalRecordResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
...
    EncounterId:
      name: encounterId
      in: path
      required: true
      schema: { type: string, format: uuid }
    EncounterServiceId:
      name: serviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ClinicalIfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag containing the Base64 encounter rowversion.
      schema:
        type: string
        pattern: '^"[A-Za-z0-9+/]{11}="$'
    BranchPublicIdQuery:
      name: branchPublicId
      in: query
      required: true
      schema:
        type: string
        pattern: '^[0-9A-Fa-f]{8}-(?:[0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication data is missing, invalid, expired, reused, or revoked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: The account is unavailable for login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: The requested resource does not exist in the caller's scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: The operation conflicts with uniqueness, concurrency, or account safety rules.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    PreconditionRequired:
      description: The If-Match precondition is required for this update.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
...
    ResponseMeta:
      type: object
      additionalProperties: true
    RequestId:
      type: string
      format: uuid
...
    ReceptionBranch:
      type: object
      additionalProperties: false
      required: [publicId, code, name, timezoneName, medicalLicenseNo, phone, email, addressLine, ward, district, province]
      properties:
        publicId: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        timezoneName: { type: string }
        medicalLicenseNo: { type: [string, 'null'] }
        phone: { type: [string, 'null'] }
        email: { type: [string, 'null'], format: email }
        addressLine: { type: string }
        ward: { type: [string, 'null'] }
        district: { type: [string, 'null'] }
        province: { type: [string, 'null'] }
...
    ClinicalBranchListResponse:
      type: object
      required: [data, meta, requestId]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ReceptionBranch' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    ClinicalEncounterStatus:
      type: string
      enum: [WAITING, IN_PROGRESS, COMPLETED, SIGNED, CANCELLED]
    ClinicalEncounterSummary:
      type: object
      required: [publicId, code, source, status, arrivedAtUtc, startedAtUtc, completedAtUtc,
        chiefComplaint, patient, doctor, room, queue]
      properties:
        publicId: { type: string, format: uuid }
        code: { type: string }
        source: { type: string, enum: [APPOINTMENT, WALK_IN] }
        status: { $ref: '#/components/schemas/ClinicalEncounterStatus' }
        arrivedAtUtc: { type: string, format: date-time }
        startedAtUtc: { type: [string, 'null'], format: date-time }
        completedAtUtc: { type: [string, 'null'], format: date-time }
        chiefComplaint: { type: [string, 'null'] }
        patient:
          type: object
          required: [publicId, code, fullName, dateOfBirth, gender]
          properties:
            publicId: { type: string, format: uuid }
            code: { type: string }
            fullName: { type: string }
            dateOfBirth: { type: string, format: date }
            gender: { type: string, enum: [MALE, FEMALE, OTHER] }
        doctor:
          type: object
          required: [publicId, fullName]
          properties:
            publicId: { type: string, format: uuid }
            fullName: { type: string }
        room:
          oneOf:
            - type: object
              required: [publicId, name]
              properties:
                publicId: { type: string, format: uuid }
                name: { type: string }
            - type: 'null'
        queue:
          oneOf:
            - type: object
              required: [displayNumber, status]
              properties:
                displayNumber: { type: string }
                status: { type: string }
            - type: 'null'
    ClinicalEncounterListResponse:
      type: object
      required: [data, meta, requestId]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/ClinicalEncounterSummary' } }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    ClinicalNotesRequest:
      type: object
      properties:
        historyOfPresentIllness: { type: [string, 'null'], maxLength: 10000 }
        physicalExamination: { type: [string, 'null'], maxLength: 10000 }
        clinicalAssessment: { type: [string, 'null'], maxLength: 10000 }
        treatmentPlan: { type: [string, 'null'], maxLength: 10000 }
        followUpInstructions: { type: [string, 'null'], maxLength: 10000 }
        followUpDate: { type: [string, 'null'], format: date }
    ClinicalStartRequest:
      type: object
      properties:
        roomPublicId: { type: [string, 'null'], format: uuid }
        queueBypassReason:
          type: [string, 'null']
          minLength: 10
          maxLength: 500
          description: Required with ENCOUNTERS_QUEUE_BYPASS when the next eligible ticket is still WAITING.
    ClinicalVitalSignsRequest:
      type: object
      properties:
        temperatureC: { type: [number, 'null'], minimum: 25, maximum: 45 }
        pulseBpm: { type: [integer, 'null'], minimum: 10, maximum: 300 }
        respiratoryRateBpm: { type: [integer, 'null'], minimum: 1, maximum: 100 }
        systolicBpMmhg: { type: [integer, 'null'], minimum: 30, maximum: 300 }
        diastolicBpMmhg: { type: [integer, 'null'], minimum: 20, maximum: 200 }
        spo2Percent: { type: [number, 'null'], minimum: 0, maximum: 100 }
        heightCm: { type: [number, 'null'], minimum: 20, maximum: 300 }
        weightKg: { type: [number, 'null'], minimum: 0.2, maximum: 500 }
        painScore: { type: [integer, 'null'], minimum: 0, maximum: 10 }
        notes: { type: [string, 'null'], maxLength: 500 }
    ClinicalDiagnosisRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string, minLength: 1, maxLength: 30 }
        name: { type: string, minLength: 2, maxLength: 500 }
        type: { type: string, enum: [PROVISIONAL, DIFFERENTIAL, FINAL], default: FINAL }
        isPrimary: { type: boolean, default: false }
        notes: { type: [string, 'null'], maxLength: 1000 }
    ClinicalOrderServiceRequest:
      type: object
      required: [servicePublicId]
      properties:
        servicePublicId: { type: string, format: uuid }
        quantity: { type: number, exclusiveMinimum: 0, maximum: 1000, default: 1 }
        notes: { type: [string, 'null'], maxLength: 1000 }
    ClinicalFinalizeResultRequest:
      type: object
      properties:
        summary: { type: [string, 'null'], maxLength: 10000 }
        conclusion: { type: [string, 'null'], maxLength: 10000 }
        result:
          oneOf:
            - type: object
              additionalProperties: true
            - type: 'null'
    ClinicalAmendmentRequest:
      type: object
      required: [reason, content]
      properties:
        reason: { type: string, minLength: 10, maxLength: 1000 }
        content: { type: string, minLength: 1, maxLength: 20000 }
...
    ClinicalEncounterResponse:
      type: object
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/ClinicalEncounterDetail' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    ClinicalCommandResponse:
      type: object
      required: [data, meta, requestId]
      properties:
        data:
          type: object
          required: [publicId]
          properties:
            publicId: { type: string, format: uuid }
            sha256: { type: string, pattern: '^[0-9A-F]{64}$' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    PatientClinicalRecordSummary:
      type: object
      additionalProperties: false
      required: [publicId, code, arrivedAtUtc, completedAtUtc, signedAtUtc, releasedAtUtc,
        chiefComplaint, patient, branch, doctor, primaryDiagnosis]
      properties:
        publicId: { type: string, format: uuid }
        code: { type: string }
        arrivedAtUtc: { type: string, format: date-time }
        completedAtUtc: { type: string, format: date-time }
        signedAtUtc: { type: string, format: date-time }
        releasedAtUtc: { type: string, format: date-time }
        chiefComplaint: { type: [string, 'null'] }
        patient:
          type: object
          additionalProperties: false
          required: [publicId, code, fullName]
          properties:
            publicId: { type: string, format: uuid }
            code: { type: string }
            fullName: { type: string }
        branch:
          type: object
          additionalProperties: false
          required: [publicId, name, timezoneName]
          properties:
            publicId: { type: string, format: uuid }
            name: { type: string }
            timezoneName: { type: string }
        doctor:
          type: object
          additionalProperties: false
          required: [publicId, fullName]
          properties:
            publicId: { type: string, format: uuid }
            fullName: { type: string }
        primaryDiagnosis:
          oneOf:
            - type: object
              additionalProperties: false
              required: [code, name]
              properties:
                code: { type: string }
                name: { type: string }
            - type: 'null'
...
    PatientClinicalRecordListResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/PatientClinicalRecordSummary' } }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
    PatientClinicalRecordResponse:
      type: object
      additionalProperties: false
      required: [data, meta, requestId]
      properties:
        data: { $ref: '#/components/schemas/PatientClinicalRecord' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        requestId: { $ref: '#/components/schemas/RequestId' }
...
    ApiErrorResponse:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
        requestId:
          $ref: '#/components/schemas/RequestId'
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true